@@ -1,89 +1,83 @@
 {
-  "name": "core-games-ui",
-  "version": "3.0.0",
+  "name": "jtb-core-games-ui",
+  "version": "0.1.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/jtbdevelopment/core-games-ui"
+  },
   "author": {
     "name": "JTB",
     "email": "jtbdevelopment@gmail.com"
   },
-  "dependencies": {
-    "@angular/common": "^4.4.5",
-    "@angular/compiler": "^4.4.5",
-    "@angular/core": "^4.4.5",
-    "@angular/http": "^4.4.5",
-    "@angular/platform-browser": "^4.4.5",
-    "@angular/platform-browser-dynamic": "^4.4.5",
-    "@angular/router": "^4.4.5",
-    "atmosphere.js": "^2.3.3",
-    "core-js": "^2.4.1",
-    "rxjs": "5.1.1",
-    "zone.js": "^0.8.4"
-  },
-  "devDependencies": {
-    "@types/es6-shim": "^0.31.32",
-    "@types/jasmine": "^2.5.43",
-    "@types/node": "^7.0.5",
-    "autoprefixer": "^6.7.3",
-    "babel-core": "^6.24.0",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.3.2",
-    "css-loader": "^0.26.1",
-    "del": "^2.2.2",
-    "es6-shim": "^0.35.3",
-    "eslint": "^3.15.0",
-    "eslint-config-xo-space": "^0.15.0",
-    "eslint-loader": "^1.6.1",
-    "eslint-plugin-babel": "^4.0.1",
-    "extract-text-webpack-plugin": "^2.0.0-rc.3",
-    "file-loader": "^0.10.1",
-    "gulp": "gulpjs/gulp#4ed9a4a3275559c73a396eff7e1fde3824951ebb",
-    "gulp-filter": "^5.0.0",
-    "gulp-hub": "frankwallis/gulp-hub#d461b9c700df9010d0a8694e4af1fb96d9f38bf4",
-    "gulp-sass": "^3.1.0",
-    "gulp-util": "^3.0.8",
-    "html-loader": "^0.4.4",
-    "html-webpack-plugin": "^2.28.0",
-    "http-proxy-middleware": "^0.17.4",
-    "jasmine": "^2.5.3",
-    "json-loader": "^0.5.4",
-    "karma": "^1.4.1",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.1.1",
-    "karma-jasmine": "^1.1.0",
-    "karma-junit-reporter": "^1.2.0",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "karma-remap-istanbul": "^0.6.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-typescript": "^3.0.0",
-    "karma-webpack": "^2.0.2",
-    "node-sass": "^4.5.0",
-    "postcss-loader": "^1.3.1",
-    "sass-loader": "^6.0.1",
-    "sourcemap-istanbul-instrumenter-loader": "^0.2.0",
-    "style-loader": "^0.13.1",
-    "ts-loader": "^2.0.0",
-    "tslint": "^4.4.2",
-    "tslint-loader": "^3.4.2",
-    "typescript": "2.3.4",
-    "url-loader": "^0.5.8",
-    "watch-ignore-webpack-plugin": "^1.0.0",
-    "webpack": "^2.2.1",
-    "webpack-fail-plugin": "^1.0.5",
-    "webpack-split-chunks": "^0.2.0"
+  "keywords": [
+    "angular",
+    "games",
+    "facebook",
+    "websocket"
+  ],
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/jtbdevelopment/core-games-ui/issues"
   },
   "scripts": {
-    "build": "gulp",
-    "test": "gulp test",
-    "test:auto": "gulp test:auto"
+    "build": "gulp build",
+    "build:watch": "gulp",
+    "docs": "npm run docs:build",
+    "docs:build": "compodoc -p tsconfig.json -n sample -d docs --hideGenerator",
+    "docs:serve": "npm run docs:build -- -s",
+    "docs:watch": "npm run docs:build -- -s -w",
+    "lint": "tslint --type-check --project tsconfig.json src/**/*.ts",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:ci": "jest --runInBand"
   },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "browser": true,
-      "jasmine": true
-    },
-    "parser": "babel-eslint",
-    "extends": [
-      "xo-space/esnext"
+  "devDependencies": {
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/compiler-cli": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/router": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@compodoc/compodoc": "^1.0.0-beta.10",
+    "@types/node": "~6.0.60",
+    "@types/jest": "^22.0.1",
+    "angular-in-memory-web-api": "^0.3.2",
+    "atmosphere.js": "^2.3.3",
+    "codelyzer": "~3.2.0",
+    "concurrently": "^3.4.0",
+    "core-js": "^2.4.1",
+    "fs-extra": "^5.0.0",
+    "del": "^2.2.2",
+    "gulp": "^3.9.1",
+    "gulp-rename": "^1.2.2",
+    "gulp-rollup": "^2.11.0",
+    "jest": "^22.1.4",
+    "jest-preset-angular": "^5.0.0",
+    "node-sass": "^4.5.2",
+    "node-sass-tilde-importer": "^1.0.0",
+    "node-watch": "^0.5.2",
+    "protractor": "~5.1.2",
+    "rollup": "^0.49.3",
+    "run-sequence": "^1.2.2",
+    "rxjs": "^5.5.2",
+    "systemjs": "^0.20.12",
+    "ts-node": "~3.2.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.4.2",
+    "zone.js": "^0.8.14"
+  },
+  "jest": {
+    "preset": "jest-preset-angular",
+    "setupTestFrameworkScriptFile": "./jest.ts",
+    "collectCoverage": true,
+    "coveragePathIgnorePatterns": [
+      "/node_modules/",
+      "jest.*"
     ]
+  },
+  "engines": {
+    "node": ">=6.0.0"
   }
 }
