@@ -2,8 +2,6 @@ import {FacebookLoginService} from './facebook-login.service';
 import {fakeAsync, TestBed, tick} from '@angular/core/testing';
 import {FacebookInitializerService} from './facebook-initializer.service';
 
-declare let window: any;
-
 class MockInitService {
     public fbReady: Promise<any>;
     public fbRequiredPermissions: string[];
@@ -41,7 +39,7 @@ describe('Service: facebook login service', () => {
         loginService.currentAuthorization.subscribe((ca) => {
             currentAuth = ca;
         });
-        window.FB = {};
+        global.FB = {};
     });
 
     it('initializes before init ready', () => {
@@ -52,7 +50,7 @@ describe('Service: facebook login service', () => {
     describe('after init is ready', () => {
         it('not connected', fakeAsync(() => {
             let called: boolean = false;
-            window.FB.getLoginStatus = (callback: (response: any) => void) => {
+            global.FB.getLoginStatus = (callback: (response: any) => void) => {
                 called = true;
                 callback({status: 'not connected'});
             };
@@ -66,10 +64,10 @@ describe('Service: facebook login service', () => {
         it('connected, missing perms', fakeAsync(() => {
             mockInit.fbRequiredPermissions = ['perm1', 'perm2', 'perm3'];
             let called: boolean = false;
-            window.FB.getLoginStatus = (callback: (response: any) => void) => {
+            global.FB.getLoginStatus = (callback: (response: any) => void) => {
                 callback({status: 'connected', authResponse: {code: 'x'}});
             };
-            window.FB.api = (api: string, callback: (response: any) => void) => {
+            global.FB.api = (api: string, callback: (response: any) => void) => {
                 called = true;
                 expect(api).toEqual('/me/permissions');
                 callback({
@@ -90,10 +88,10 @@ describe('Service: facebook login service', () => {
         it('connected, have perms', fakeAsync(() => {
             mockInit.fbRequiredPermissions = ['perm1', 'perm2', 'perm3'];
             let called: boolean = false;
-            window.FB.getLoginStatus = (callback: (response: any) => void) => {
+            global.FB.getLoginStatus = (callback: (response: any) => void) => {
                 callback({status: 'connected', authResponse: {x: 1, y: '32'}});
             };
-            window.FB.api = (api: string, callback: (response: any) => void) => {
+            global.FB.api = (api: string, callback: (response: any) => void) => {
                 called = true;
                 expect(api).toEqual('/me/permissions');
                 callback({
@@ -115,32 +113,32 @@ describe('Service: facebook login service', () => {
 
     it('manually initiating login before init does nothing', () => {
         mockInit.fbRequiredPermissions = ['perm1', 'perm3'];
-        window.FB.login = jasmine.createSpy('login');
+        global.FB.login = jest.fn();
         loginService.initiateLogin();
-        expect(window.FB.login).not.toHaveBeenCalled();
+        expect(global.FB.login.mock.calls.length).toEqual(0);
     });
 
     it('manually initiating login after init', fakeAsync(() => {
         mockInit.fbRequiredPermissions = ['perm1', 'perm3'];
 
         //  initially not connected
-        window.FB.getLoginStatus = (callback: (response: any) => void) => {
+        global.FB.getLoginStatus = (callback: (response: any) => void) => {
             callback({status: 'not connected'});
         };
         mockInit.resolve(true);
         tick();
 
 
-        window.FB.login = (callback: () => void, data: any) => {
+        global.FB.login = (callback: () => void, data: any) => {
             expect(data).toEqual({scope: 'perm1,perm3'});
             callback();
         };
 
         let called: boolean = false;
-        window.FB.getLoginStatus = (callback: (response: any) => void) => {
+        global.FB.getLoginStatus = (callback: (response: any) => void) => {
             callback({status: 'connected', authResponse: {x: 2, y: '32'}});
         };
-        window.FB.api = (api: string, callback: (response: any) => void) => {
+        global.FB.api = (api: string, callback: (response: any) => void) => {
             called = true;
             expect(api).toEqual('/me/permissions');
             callback({
