@@ -7,9 +7,9 @@ export abstract class AbstractTurnClassifier {
     protected static THEIR_TURN: string = 'Their Turn';
     protected static OLDER_GAMES: string = 'Older Games';
 
-    private static DEFAULT_CLASSIFICATIONS: BehaviorSubject<string[]> = new BehaviorSubject<string[]>(['Your Turn', 'Their Turn', 'Older Games']);
+    private DEFAULT_CLASSIFICATIONS: BehaviorSubject<string[]> = new BehaviorSubject<string[]>(['Your Turn', 'Their Turn', 'Older Games']);
 
-    private static DEFAULT_ICONS: BehaviorSubject<Map<string, string>> = new BehaviorSubject<Map<string, string>>(new Map<string, string>(
+    private DEFAULT_ICONS: BehaviorSubject<Map<string, string>> = new BehaviorSubject<Map<string, string>>(new Map<string, string>(
         [
             [AbstractTurnClassifier.YOUR_TURN, 'play'],
             [AbstractTurnClassifier.THEIR_TURN, 'pause'],
@@ -19,12 +19,12 @@ export abstract class AbstractTurnClassifier {
 
     //noinspection JSMethodCanBeStatic
     public getClassifications(): Observable<string[]> {
-        return Observable.from(AbstractTurnClassifier.DEFAULT_CLASSIFICATIONS);
+        return Observable.from(this.DEFAULT_CLASSIFICATIONS);
     }
 
     //  Game buckets to icons - can return empty map if not initialized
     //noinspection JSMethodCanBeStatic
     public getIcons(): Observable<Map<string, string>> {
-        return Observable.from(AbstractTurnClassifier.DEFAULT_ICONS);
+        return Observable.from(this.DEFAULT_ICONS);
     }
 }
